Organize the files into packages
