+++ watershed-android/watershed/app/src/main/AndroidManifest.xml
@@ -8,7 +8,7 @@
         android:label="@string/app_name"
         android:theme="@style/Theme.AppCompat.Light" >
         <activity
-            android:name=".LandingPageActivity"
+            android:name=".Activities.LandingPageActivity"
             android:label="@string/title_activity_landing_page"
             android:theme="@style/AppThemeTranslucent"
             android:screenOrientation="portrait">
@@ -19,7 +19,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".Activities.MainActivity"
             android:label="@string/title_activity_main"
             android:screenOrientation="portrait">
         </activity>
