+++ watershed-android/watershed/app/src/main/java/com/blueprint/watershed/LoginFragment.java
@@ -2,26 +2,15 @@ package com.blueprint.watershed;
 
 import android.app.Activity;
 import android.content.SharedPreferences;
-import android.net.Uri;
 import android.os.Bundle;
 import android.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.EditText;
 import android.widget.Button;
-import android.content.Intent;
-import android.widget.Toast;
 
-import com.android.volley.NetworkError;
-import com.android.volley.Request;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.JsonObjectRequest;
-
-import org.json.JSONException;
-import org.json.JSONObject;
+import com.blueprint.watershed.Activities.LandingPageActivity;
 
 import java.util.HashMap;
