+++ watershed-android/watershed/app/src/main/java/com/blueprint/watershed/Networking/RequestHandler.java
@@ -1,7 +1,6 @@
-package com.blueprint.watershed;
+package com.blueprint.watershed.Networking;
 
 import android.content.Context;
-import android.content.SharedPreferences;
 import android.graphics.Bitmap;
 import android.support.v4.util.LruCache;
 import android.util.Log;
@@ -19,8 +18,6 @@ import com.android.volley.toolbox.Volley;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
-
-import java.util.ArrayList;
 
 /*
 Singleton Request Handler to interface with Network.
