+++ watershed-android/watershed/app/src/main/java/com/blueprint/watershed/Sites/Site.java
@@ -1,4 +1,4 @@
-package com.blueprint.watershed;
+package com.blueprint.watershed.Sites;
 
 /**
  * Created by Mark Miyashita on 10/14/14.
