+++ watershed-android/watershed/app/src/main/java/com/blueprint/watershed/Tasks/Task.java
@@ -1,4 +1,4 @@
-package com.blueprint.watershed;
+package com.blueprint.watershed.Tasks;
 
 import android.util.Log;
