+++ watershed-android/watershed/app/src/main/java/com/blueprint/watershed/Tasks/TaskAdapter.java
@@ -1,4 +1,4 @@
-package com.blueprint.watershed;
+package com.blueprint.watershed.Tasks;
 
 import android.app.Activity;
 import android.content.Context;
@@ -7,6 +7,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
+
+import com.blueprint.watershed.R;
 
 import java.util.ArrayList;
