+++ watershed-android/watershed/app/src/main/java/com/blueprint/watershed/Tasks/TaskDetailFragment.java
@@ -1,14 +1,15 @@
-package com.blueprint.watershed;
+package com.blueprint.watershed.Tasks;
 
 import android.app.Activity;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import com.blueprint.watershed.R;
 
 import java.util.Date;
