+++ watershed-android/watershed/app/src/main/java/com/blueprint/watershed/Tasks/TaskFragment.java
@@ -1,21 +1,21 @@
-package com.blueprint.watershed;
+package com.blueprint.watershed.Tasks;
 
 import android.app.Activity;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.support.v4.app.ListFragment;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.ListAdapter;
 import android.widget.ListView;
-import android.widget.Toast;
 
 import com.android.volley.Request;
 import com.android.volley.Response;
+import com.blueprint.watershed.Activities.MainActivity;
+import com.blueprint.watershed.Networking.BaseRequest;
+import com.blueprint.watershed.R;
+import com.blueprint.watershed.Networking.RequestHandler;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
@@ -23,7 +23,6 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.HashMap;
 
 public class TaskFragment extends ListFragment {
