+++ watershed-android/watershed/app/src/main/java/com/blueprint/watershed/Views/CoverPhotoView.java
@@ -1,4 +1,4 @@
-package com.blueprint.watershed;
+package com.blueprint.watershed.Views;
 
 import android.content.Context;
 import android.util.AttributeSet;
@@ -7,6 +7,8 @@ import android.view.LayoutInflater;
 import android.widget.ImageView;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
+
+import com.blueprint.watershed.R;
 
 /**
  * Created by Mark Miyashita on 10/4/14.
