+++ watershed-android/watershed/app/src/main/res/layout/fragment_site.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.blueprint.watershed.SiteFragment">
+    tools:context="com.blueprint.watershed.Sites.SiteFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
