+++ watershed-android/watershed/app/src/main/res/layout/fragment_task_detail.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.blueprint.watershed.TaskDetailFragment">
+    tools:context="com.blueprint.watershed.Tasks.TaskDetailFragment">
 
     <TextView
         android:id = "@+id/task_detail_title"
