+++ watershed-android/watershed/app/src/main/res/layout/fragment_task_grid.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.blueprint.watershed.TaskFragment">
+    tools:context="com.blueprint.watershed.Tasks.TaskFragment">
 
     <GridView
         android:id="@android:id/list"
