+++ watershed-android/watershed/app/src/main/res/layout/site_list_row.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <com.blueprint.watershed.CoverPhotoView
+    <com.blueprint.watershed.Views.CoverPhotoView
         android:layout_width="match_parent"
         android:layout_height="200dp"
         />
