+++ watershed-android/watershed/app/src/main/res/menu/main.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.blueprint.watershed.MainActivity" >
+    tools:context="com.blueprint.watershed.Activities.MainActivity" >
     <item android:id="@+id/action_settings"
         android:title="@string/action_settings"
         android:orderInCategory="100"
